fix!: changing backend API naming
BREAKING CHANGE: Some urls changed
